requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.7
django-taggit==0.22.1
django-toolbelt==0.0.1
gunicorn==19.7.1
model-mommy==1.4.0
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
six==1.11.0
static3==0.7.0
